VIC-9888 Add /etc/victor-compat-version
We want to stop victor.git developers from deploying incompatible code
onto robots.  Sometimes their code is too new for the robot and requires
a feature or bug fix that the robot doesn't have.  Sometimes, it is the
other way around and a change has been made in VicOS that requires a
corresponding change in victor.git for everything to work together
correctly.  A victor.git developer on an older branch may not know this,
try to deploy, and experience negative results.

In the past, we tried to do this by changing the os version.  Like, if
we were at version 1.2.0 and introduced an incompatible change, we would
increment it to 1.2.1.  That worked, but caused trouble for the release
engineering team and the product management team who made plans around
what os version number would be released to end customers.

This new version, /etc/victor-compat-version, is for our internal use
only and will not be displayed to end users.  A corresponding change
will be made in victor.git to check this version before deploying.