VIC-9974 Allow vic-robot to read EMR (#623)
In order to support Whiskey, vic-robot needs to be able to read the EMR on startup. Previously vic-robot was relying on the vic-engine.service to set the anki group on the emr which was fine as vic-robot did not try to read from the emr until it had established comms with vic-engine.