VIC-12781 Enable network-online.target via connman
We want to try and download an OTA as soon as we first gain network
access.  This change adds the `connman-wait-online` service.
`systemd` uses this service to determine if the
`network-online.target` is active and therefore it can activate units
that depend on `network-online.target`.  A later commit will add a
service that waits for `network-online.target` and then immediately
tries to do an OTA update.