VIC-12781 Check for OTA as soon as robot is online
We are concerned that some users are turning their robots on, playing
with them for a little while, and then turning them off before they
can download and install the latest OTA.  This happens because we are
not terribly aggressive about trying to download an OTA file.
Instead, once an hour, we delay a random amount of time between 1
second and 1 hour and then try to download the OTA file.  We do this
delay so that we don't have all of the robots hitting the OTA server
at the exact same time.  As a result, you could have a robot that was
powered on at 9:05am, and won't try to update until sometime after
10am.  If the user turns off their robot before that happens, they
will not get the update.

With this change, as soon as the robot gets on WiFi it will try to
download an OTA update.