VIC-13784 Report OS version to boot log
When diagnosing a bug on the robot, we will often look at
`/data/boot.log`.  It is helpful to know what OS version the robot is
running.  With this change, we will report the OS version to `/dev/kmsg`
which will make it available in `/data/boot.log` and when running
`dmesg`.