VIC-13882 viccubetool - connection stress test
A stress test mode for connecting to Vector's cube.  In a loop, the test
will scan for cubes, connect to the first cube found (or the one given
on the command line as an argument), handle any received information,
idle out and disconnect.  At the end of the requested iterations, some
basic summary statistics will be printed out.

Usage Examples:
viccubetool conntest 10
viccubetool -a c2:40:b1:98:26:ff conntest 5