VIC-12404 Set ANKI_VERSION to 0.16.0 for RMA build
This build will be used by Customer Care / Engineering / Etc. to
diagnose returned robots.  The build will self destruct and will not be
available to the general public.

To note, this change updates the `anki/victor` submodule pointer to pick
up the corresponding change to the `VERSION` file in that repository.