VIC-12404 Self Destruct Build
This build will be used by Customer Care to diagnose returned robots.
After it is installed on the robot, it should only be used one time and
then self destruct.  The next boot should return to the recovery
slot (aka the `f` slot).  In addition, we want to ensure that the
customer's data on the switchboard and data partitions is erased.