VIC-12404 Disable automatic reboots
After we install the RMA build on a returned robot, we do not want it to
do the automatic overnight reboot.  The customer care agent, engineer,
etc. will reboot the robot when they are done diagnosing it.  This RMA
build is self-destructing.  A reboot is a one way trip to the recovery
partition with the userdata and switchboard partitions getting wiped out.