VIC-8375 Backport fix for modem wait state for APQ8009 (#654)
Qualcomm fixed a bug in the 4.9 kernel related to the modem wait state on apq8009
boards. Since the apq8009 does not have a cell modem, the modem state moves quickly
from DOWN -> UP, without a long "LOADING" phase. The original logic to retry the probe if
`state != APR_SUBSYS_LOADED` only works properly on msm8909 boards, and could fail
in our case if the state is `APR_SUBSYS_UP` when the modem firmware is loaded
 or the audio codec is probed. Checking `state == APR_SUBSYS_DOWN` should work for
both boards.

Related: VIC-13874

This is the commit from the 4.9 kernel
https://source.codeaurora.org/quic/la/platform/vendor/opensource/audio-kernel/commit/?h=LE.UM.2.3.1.r1.2&id=3745fe668dd2edb0c5c881a4151df11d6920ca0f

```
commit 3745fe668dd2edb0c5c881a4151df11d6920ca0f
Author: yang liu <yangliu@codeaurora.org>
Date:   Fri Aug 10 12:30:01 2018 +0800

    ASoC: codec: add wait modem state for APQ8009

    On msm platform, internal codecs wait for adsp loaded. On apq8009,
    internal codecs need wait for modem up. Current logic only works
    for msm device, not apq devices. Update analog and digital codec
    probe functions to fit for both msm and apq platforms.

    Change-Id: Ib2c848b7aaddef18cee3f658fe895d436762b5f0
    Signed-off-by: yang liu <yangliu@codeaurora.org>
```