VIC-14140 Smoother Delta OTA Progress Reporting
As noted in the corresponding commit for victor.git, we were only
reporting progress for a small fraction of the delta OTA operations.  We
were omitting reporting progress on calculating the SHA256 hash of the
system and boot partitions and the copy of the old partitions to the new
partitions.  This is a significant amount of time for applying a delta OTA.
With this change, we report progress much more frequently.  In addition,
this change will report back the current phase of operation so that
`update-engine` can write it out to `/run/update-engine/phase`.

This change requires a complementary change in victor.git and thus
VICTOR_COMPAT_VERSION is updated here.