VIC-11720 Restrict /data/lib/connman to root
Previously, we had changed the owner of /data/lib/connman to be
net:anki.  We did this so that vic-switchboard could deposit a
wifi.settings file in that directory and have connman 1.30 pick it up
and use it to configure wifi on the robot.  With connman 1.36, that
approach doesn't work.  Instead, in vic-switchboard, we have to
provide an Agent interface to feed the passphrase to connman when
prompted.  connman will then create its own subdirectory below
/data/lib/connman to store the info about the wifi network.  However,
it will fail to do so if /data/lib/connman is not owned by root with
restrictive permissions.  Presumably this is a security safeguard to
prevent non-root processes from reading privileged information (like
the WiFi passphrase).  With this change, we make sure that
/data/lib/connman is owned by root.