VIC-7566 put /dev/ion under control of new 'ion' group.
/dev/ion was owned by the camera group. A number of processes will
need to be able to access this device to be able to access shared
memory, not all of them camera-related.

This creates an 'ion' group and makes it the owner of /dev/ion. It
also adds the robot and engine users to the new group, in preparation
for vic-engine and vic-robot needing to access a new shared memory
buffer.