VIC-14343: Move custom layers out of poky directory
This commit moves all custom layers out of the poky directory so yocto can
be updated independent of vendor releases.

All of the Qualcomm layers for victor have been moved to the new
meta-victor-bsp layer. The meta-victor-bsp layer is indended to include
all of the victor platform specific layers and metadata.

The layers that are indended for sharing across platforms are under the
top level oe directory.