VIC-14404 update-engine : track sleeping state
Every hour, we do the periodic OTA check.  We first sleep (aka delay)
for a random amount of time between 1 second and 1 hour.  We do this so
that all the robots are not hitting the OTA server at the same time.
However, this means that if the user wants to manually initiate an OTA
update from Chewie, they will sit there waiting and waiting for the
sleep to finish.  With this change, we will record the "sleeping" phase
in the `/run/update-engine/phase` file so that `vic-gateway` can detect
it and restart the OTA client so that it will update immediately.

In addition, this change handles the rare case were `update-engine` is
killed and doesn't get to write out `/run/update-engine/exit_code`.