SEC-1 patch kernel: CVE-2018-10879 (1/2)
A flaw was found in the Linux kernel's ext4 filesystem.
A local user can cause a use-after-free in ext4_xattr_set_entry
function and a denial of service or unspecified other
impact may occur by renaming a file in a crafted ext4 filesystem image.

This is the first of 2 patches

This is the commit from the mainline:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=513f86d73855ce556ea9522b6bfd79f87356dc3a
```
author  Theodore Ts'o <tytso@mit.edu>   2018-06-13 00:51:28 -0400
commit  513f86d73855ce556ea9522b6bfd79f87356dc3a

ext4: always verify the magic number in xattr blocks
If there an inode points to a block which is also some other type of
metadata block (such as a block allocation bitmap), the
buffer_verified flag can be set when it was validated as that other
metadata block type; however, it would make a really terrible external
attribute block.  The reason why we use the verified flag is to avoid
constantly reverifying the block.  However, it doesn't take much
overhead to make sure the magic number of the xattr block is correct,
and this will avoid potential crashes.

This addresses CVE-2018-10879.

https://bugzilla.kernel.org/show_bug.cgi?id=200001

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
```