SEC-1 patch kernel: CVE-2018-10880 (1/1)
Linux kernel is vulnerable to a stack-out-of-bounds write
in the ext4 filesystem code when mounting and writing to a
crafted ext4 image in ext4_update_inline_data(). An attacker
could use this to cause a system crash and a denial of service.

This is the commit from the mainline:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=8cdb5240ec5928b20490a2bb34cb87e9a5f40226
```
author  Theodore Ts'o <tytso@mit.edu>   2018-06-16 15:40:48 -0400
commit  8cdb5240ec5928b20490a2bb34cb87e9a5f40226

ext4: never move the system.data xattr out of the inode body
When expanding the extra isize space, we must never move the
system.data xattr out of the inode body.  For performance reasons, it
doesn't make any sense, and the inline data implementation assumes
that system.data xattr is never in the external xattr block.

This addresses CVE-2018-10880

https://bugzilla.kernel.org/show_bug.cgi?id=200005

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
```