SEC-1 patch kernel: CVE-2018-10883 (2/2)
A flaw was found in the Linux kernel's ext4 filesystem.
A local user can cause an out-of-bounds write in
jbd2_journal_dirty_metadata(), a denial of service,
and a system crash by mounting and operating on a crafted
ext4 filesystem image.

This is the second of 2 related patches

This is the commit from the mainline:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=8bc1379b82b8e809eef77a9fedbb75c6c297be19
```
author  Theodore Ts'o <tytso@mit.edu>   2018-06-16 23:41:59 -0400
commit  8bc1379b82b8e809eef77a9fedbb75c6c297be19

ext4: avoid running out of journal credits when appending to an inline file
Use a separate journal transaction if it turns out that we need to
convert an inline file to use a data block.  Otherwise, we could end
up failing due to not having journal credits.

This addresses CVE-2018-10883.

https://bugzilla.kernel.org/show_bug.cgi?id=200071

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
```