SEC-1 patch kernel: CVE-2018-10883 (1/2)
A flaw was found in the Linux kernel's ext4 filesystem.
A local user can cause an out-of-bounds write in
jbd2_journal_dirty_metadata(), a denial of service,
and a system crash by mounting and operating on a crafted
ext4 filesystem image.

This is the first of 2 related patches

This is the commit from the mainline:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=e09463f220ca9a1a1ecfda84fcda658f99a1f12a
```
author  Theodore Ts'o <tytso@mit.edu>   2018-06-16 20:21:45 -0400
commit  e09463f220ca9a1a1ecfda84fcda658f99a1f12a

jbd2: don't mark block as modified if the handle is out of credits
Do not set the b_modified flag in block's journal head should not
until after we're sure that jbd2_journal_dirty_metadat() will not
abort with an error due to there not being enough space reserved in
the jbd2 handle.

Otherwise, future attempts to modify the buffer may lead a large
number of spurious errors and warnings.

This addresses CVE-2018-10883.

https://bugzilla.kernel.org/show_bug.cgi?id=200071

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
```